+++ broker/src/main/resources/db/migration/V1_0_22__add_service_context_to_updateRequest.sql
@@ -1,0 +1,6 @@
+ALTER TABLE update_request
+  ADD COLUMN service_context_id BIGINT;
+
+ALTER TABLE update_request
+  ADD CONSTRAINT FK_update_request2service_context
+FOREIGN KEY (service_context_id) REFERENCES service_context (id);